update interest and delete interest functionality added

+++ Services/Sober/Sober.Application/Pages/Educations/Commands/DeleteEducationCommandHandler.cs
@@ -27,7 +27,7 @@ public class DeleteEducationCommandHandler
         }
 
         bool isDeleted = _educationRepository.DeleteEducation(request.educationId);
-        if (isDeleted)
+        if (isDeleted && education.InstituteLogo is not null)
         {
             _fileService.DeleteFileAsync(education.InstituteLogo);
         }
